# ai-specs/specs/api-spec.yml
openapi: 3.1.0
info:
  title: ADA ASM API
  version: 0.1.0
  description: |
    REST API for ADA ASM. This file is the source of truth for the API
    contract and is regenerated from the FastAPI app's `/openapi.json`
    on every PR that changes `backend/app/api/**`.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/modules:
    get:
      tags: [modules]
      summary: List modules (catalogue), aggregates hydrated server-side
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: ILIKE match against lower(sku), lower(name), lower(description)
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Paginated list with each item carrying its aggregated_* fields
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [modules]
      summary: Create a new module
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Module + aggregates + children + parents
        "409":
          description: MODULE_SKU_ALREADY_REGISTERED
  /api/v1/modules/{module_id}:
    parameters:
      - in: path
        name: module_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [modules]
      summary: Module detail with hydrated aggregates, children, and parents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ModuleResponse — extends Module with children + parents arrays
        "404":
          description: MODULE_NOT_FOUND
    patch:
      tags: [modules]
      summary: Partial update; id/created_at/updated_at ignored if sent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Module re-hydrated with aggregates
        "404":
          description: MODULE_NOT_FOUND
        "409":
          description: MODULE_SKU_ALREADY_REGISTERED
    delete:
      tags: [modules]
      summary: Delete a module (CASCADE removes its edges but not its children)
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
        "404":
          description: MODULE_NOT_FOUND
  /api/v1/modules/{module_id}/children:
    parameters:
      - in: path
        name: module_id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [modules]
      summary: Add a child (component XOR module) with quantity
      description: |
        Validates XOR + existence + cycle detection (WITH RECURSIVE) +
        duplicate pair. Returns the hydrated edge with `child_module` or
        `child_component` summary populated.
      security:
        - bearerAuth: []
      responses:
        "201":
          description: ModuleChildResponse with hydrated summary
        "404":
          description: MODULE_NOT_FOUND (parent missing)
        "422":
          description: |
            One of: INVALID_CHILD_REFERENCE (XOR violation OR missing referenced row),
            MODULE_CYCLE_DETECTED, CHILD_ALREADY_PRESENT.
  /api/v1/modules/{module_id}/children/{child_id}:
    parameters:
      - in: path
        name: module_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: child_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [modules]
      summary: Update quantity / notes / sort_order of an existing edge
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated ModuleChildResponse
        "404":
          description: MODULE_NOT_FOUND (parent or edge missing)
        "422":
          description: quantity < 1
    delete:
      tags: [modules]
      summary: Remove an edge — idempotent on missing
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Edge removed (or never existed)
  /api/v1/modules/{module_id}/price-history:
    parameters:
      - in: path
        name: module_id
        required: true
        schema: { type: string, format: uuid }
      - in: query
        name: period
        schema: { type: string, enum: [week, month, year], default: year }
    get:
      tags: [modules]
      summary: Aggregated price time-series (server-computed from descendant components)
      description: |
        For each `valid_from` date in the period, sums
        `quantity_propagated × precio_vigente(componente, fecha)` over every
        descendant component using its preferred supplier at qty_tier=100.
        Empty `series` when no descendants have supplier_prices.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "ModulePriceHistoryResponse with series of {date, price} points"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModulePriceHistoryResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: MODULE_NOT_FOUND
  /api/v1/modules/{module_id}/stock-events:
    parameters:
      - in: path
        name: module_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [modules]
      summary: Module-level stock events (fabricated / delivered) ordered by `occurred_at DESC`
      description: |
        Drives the "Histórico de fabricación" modal. Returns events with
        `module_id` set and `kind in ('fabricated', 'delivered')`. Supports
        pagination.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Paginated `StockEvent` rows for this module.
          content:
            application/json:
              schema:
                type: object
                required: [items, total, page, page_size]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/StockEvent" }
                  total: { type: integer, minimum: 0 }
                  page: { type: integer, minimum: 1 }
                  page_size: { type: integer, minimum: 1 }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: MODULE_NOT_FOUND
  /api/v1/modules/{module_id}/component-purchases-summary:
    parameters:
      - in: path
        name: module_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [modules]
      summary: Aggregated supplier spend across descendant components
      description: |
        For each supplier that has been purchased from for any descendant
        component of this module, returns total qty + total cost (EUR).
        Drives the "Proveedor más comprado" pane on the module detail and
        the "Historial de fabricación" modal.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of `SupplierPurchaseSummary` rows.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SupplierPurchaseSummary" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: MODULE_NOT_FOUND
  /api/v1/modules/{module_id}/projects-using:
    parameters:
      - in: path
        name: module_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [modules]
      summary: Projects that hold this module as a direct child
      description: |
        Drives the "Usado en proyectos" section on the module detail. Returns
        hydrated ProjectSummary rows.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of `ProjectSummary` rows.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectSummary" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: MODULE_NOT_FOUND
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 Problem Details with a stable `code` discriminator.
      required: [title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
...
    StockEvent:
      type: object
      description: |
        Append-only ledger row. Polymorphic owner: exactly one of `component_id` /
        `module_id` is non-null per row (DB-level XOR CHECK).

        Kind dictates which fields are populated:
          * `purchase`     — component-level restock from a supplier.
            Fills: component_id, supplier_id, unit_cost, total_cost, currency.
          * `consumption`  — component allocated to a project.
            Fills: component_id, project_id, project_name_snapshot.
          * `fabricated`   — module assembled in-house. Adds units to the module's stock.
            Fills: module_id, unit_cost, total_cost (computed cost per assembled unit).
          * `delivered`    — module shipped to a customer. Removes units from stock.
            Fills: module_id, customer_id_holded, customer_name_snapshot.
      required: [id, kind, quantity, occurred_at, currency, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        component_id:
          type: string
          format: uuid
          nullable: true
          description: Set for component-level events (`purchase`, `consumption`). Null for module events.
        module_id:
          type: string
          format: uuid
          nullable: true
          description: Set for module-level events (`fabricated`, `delivered`). Null for component events.
        kind:
          type: string
          enum: [purchase, consumption, fabricated, delivered]
        quantity: { type: integer, minimum: 1 }
        occurred_at: { type: string, format: date }
        notes: { type: string, nullable: true }
        supplier_id: { type: string, format: uuid, nullable: true }
        supplier_name: { type: string, nullable: true, description: "JOIN with suppliers.name (server-set)" }
        unit_cost: { type: string, nullable: true, description: Decimal as string }
        total_cost: { type: string, nullable: true, description: Decimal as string }
        currency: { type: string, minLength: 3, maxLength: 3 }
        project_id: { type: string, format: uuid, nullable: true }
        project_name_snapshot: { type: string, nullable: true }
        customer_id_holded:
          type: string
          nullable: true
          maxLength: 64
          description: "`delivered` only — Holded customer ID for the module shipment."
        customer_name_snapshot:
          type: string
          nullable: true
          maxLength: 200
          description: "`delivered` only — denormalised customer display name for audit."
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ----- Modules -----
...
    ModulePriceHistoryPoint:
      type: object
      required: [date, price]
      properties:
        date: { type: string, format: date }
        price:
          type: string
          description: Decimal as string; the module's `precio_total` at that date.
    ModulePriceHistoryResponse:
      type: object
      required: [period, points]
      properties:
        period: { type: string, enum: [week, month, year] }
        points:
          type: array
          items: { $ref: "#/components/schemas/ModulePriceHistoryPoint" }
    SupplierPurchaseSummary:
      type: object
      description: |
        Aggregated supplier spend across all descendant components of a module.
        Drives the "Proveedor más comprado" pane on the module detail / historial.
      required: [supplier_name, qty, cost]
      properties:
        supplier_id: { type: string, format: uuid, nullable: true }
        supplier_name: { type: string }
        qty: { type: integer, minimum: 0 }
        cost: { type: string, description: Decimal as string (EUR). }
...
    ProjectStatus:
      type: string
      enum: [Presupuestado, Esperando, En proceso, Completado, Archivado]
      description: |
        Lifecycle status — DB-enforced (CHECK) and shared with the FE labels verbatim.
        Soft-delete transitions to `Archivado`. The PATCH endpoint auto-fills
        `fecha_entrega_real` with today's date when the status changes to
        `Completado` and the body does not provide an explicit value.
    Customer:
      type: object
      description: Id-link to Holded; the FE renders `holded_id` as a clickable badge.
      required: [id, holded_id, name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        holded_id: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 200 }
        holded_url: { type: string, nullable: true, maxLength: 500 }
        notas: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Project:
      type: object
      description: Top of the asset tree. Soft-delete only (transition to Archivado).
      required: [id, code, name, status, tags, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, maxLength: 40, description: "Surfaced in the UI as `Clave`." }
        name: { type: string, maxLength: 200 }
        description: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        customer_id: { type: string, format: uuid, nullable: true }
        icon:
          type: string
          nullable: true
          maxLength: 8
          description: Single emoji char/cluster — rendered in the list row icon cell and detail header.
        color:
          type: string
          nullable: true
          maxLength: 7
          description: "Hex `#rrggbb`. Used as visual accent (10% bg on the emoji)."
        tags:
          type: array
          items: { type: string }
          description: Free-text labels (e.g. `power`, `motor`, `automotive`).
        version: { type: string, nullable: true, maxLength: 40 }
        fecha_inicio: { type: string, format: date, nullable: true }
        fecha_entrega_estimada: { type: string, format: date, nullable: true }
        fecha_entrega_real: { type: string, format: date, nullable: true }
        notas: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProjectSummary:
      type: object
      description: Project + server-computed aggregates + embedded customer summary.
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            customer:
              allOf:
                - $ref: "#/components/schemas/Customer"
              nullable: true
            precio_total:
              type: string
              nullable: true
              description: Decimal as string (EUR).
            aggregated_nato_score:
              type: string
              enum: ["A+", A, B, C, D, F]
              nullable: true
            aggregated_tier:
              type: integer
              enum: [1, 2, 3, 4]
              nullable: true
            aggregated_expires_at:
              type: string
              format: date
              nullable: true
            buildable_stock:
              type: integer
              minimum: 0
...
  responses:
    Unauthorized:
      description: |
        Authentication failed. `code` is one of:
        `INVALID_CREDENTIALS`, `UNAUTHENTICATED`, `ACCESS_TOKEN_EXPIRED`,
        `ACCESS_TOKEN_WRONG_TYPE`, `ACCESS_TOKEN_MALFORMED`,
        `REFRESH_TOKEN_EXPIRED`, `REFRESH_TOKEN_REVOKED`,
        `REFRESH_TOKEN_MALFORMED`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...